backend/requirements.txt
django
djangorestframework
fastexcel
polars
pyyaml
requests
